requirements.txt
# CyMind 平台依赖

# Web框架
Flask==2.3.3
Flask-CORS==4.0.0

# 数据库
SQLAlchemy==2.0.21
alembic==1.12.0

# 测试框架
pytest==7.4.2
pytest-cov==4.1.0
hypothesis==6.88.1

# 配置管理
PyYAML==6.0.1

# 报告生成
Jinja2==3.1.2
WeasyPrint==60.0
markdown==3.5.1

# AI集成 (可选)
openai==0.28.1
langchain==0.0.330

# 任务队列 (可选)
celery==5.3.2
redis==5.0.1

# 安全工具集成
python-nmap==0.7.1
requests==2.31.0
beautifulsoup4==4.12.2

# 数据处理
pandas==2.1.1
numpy==1.25.2

# 网络工具
dnspython==2.4.2
ipaddress==1.0.23

# 开发工具
black==23.9.1
flake8==6.1.0
mypy==1.6.0

# 生产部署
gunicorn==21.2.0
supervisor==4.2.5

# 监控和日志
prometheus-client==0.17.1
structlog==23.1.0